% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectPie.R
\name{cdfevd}
\alias{cdfevd}
\title{cdfevd}
\usage{
cdfevd(x)
}
\arguments{
\item{x}{Numeric vector. Values to be transformed via the standardized
Gumbel CDF to the (0, 1) interval.}
}
\value{
Numeric vector of CDF values in (0, 1).
}
\description{
Marginal CDF helper function (Gumbel / extreme value distribution).
}
