% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectPie-package.R
\docType{package}
\name{SelectPie-package}
\alias{SelectPie-package}
\title{SelectPie: MLE for Compositional Models with Sample Selection}
\description{
Maximum likelihood estimation for compositional outcome models
that account for selection into the sample using a bivariate
extreme value (Gumbel-logistic) copula.
}
\author{
\strong{Maintainer}: First Last \email{you@email.com}

}
